backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
django-storages==1.14.2
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Database
psycopg2-binary==2.9.7
redis==5.0.1

# Image processing
Pillow==10.0.1

# Background tasks
celery==5.3.4
django-celery-beat==2.5.0

# Real-time features
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Payments
stripe==7.6.0

# Geo features
django-geoposition==0.3.2
geopy==2.4.0

# API documentation
drf-yasg==1.21.7

# Environment variables
python-decouple==3.8

# Email
django-anymail==10.2

# Search
django-elasticsearch-dsl==8.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0 